review: rename util.rs -> skip_partial.rs
